♻️📝 refactor and doc: added new function to create JSON, refactored code and added documentation to functions ✨ feat: Added excel file to project and change route of excel
